T-271 Moved helper to correct package, edited variables as needed
